Moved stringvalues to dialog package
