+++ src/tool/clients/fmmlxdiagrams/dialogs/ChangeMultiplicityDialog.java
@@ -17,7 +17,7 @@ import tool.clients.fmmlxdiagrams.FmmlxAttribute;
 import tool.clients.fmmlxdiagrams.FmmlxDiagram;
 import tool.clients.fmmlxdiagrams.FmmlxObject;
 import tool.clients.fmmlxdiagrams.dialogs.results.ChangeMultiplicityDialogResult;
-import tool.clients.fmmlxdiagrams.stringvalue.StringValueDialog;
+import tool.clients.fmmlxdiagrams.dialogs.stringvalue.StringValueDialog;
 
 public class ChangeMultiplicityDialog extends CustomDialog<ChangeMultiplicityDialogResult> {
